More examples

+++ src/CANopen/EMCY/EMCY_Internal/EMCY_Protocol/CANopen_EMCY_Protocol_Produce.c
@@ -36,7 +36,7 @@ void CANopen_EMCY_Protocol_Produce_Error(CANopen *canopen, uint16_t new_error_co
 	data[7] = vendor_specific_data[4];
 
 	/* Create the COB ID */
-	uint32_t COB_ID = FUNCTION_CODE_SYNC_EMCY << 7 | canopen->slave.this_node_ID;
+	uint16_t COB_ID = FUNCTION_CODE_SYNC_EMCY << 7 | canopen->slave.this_node_ID;
 
 	/* Send the message to client */
 	Hardware_CAN_Send_Message(COB_ID, data);
