+++ src/CANopen/HEARTBEAT/HEARTBEAT_Internal/HEARTBEAT_Protocol/CANopen_HEARTBEAT_Protocol_Produce.c
@@ -17,7 +17,7 @@ void CANopen_HEARTBEAT_Protocol_Produce_Status(CANopen *canopen, uint8_t node_ID
 	data[0] = (canopen->slave.nmt.this_node_toggle << 7) | canopen->slave.nmt.this_node_status_operational;
 
 	/* Create the COB ID and send heartbeat to the node */
-	uint32_t COB_ID = FUNCTION_CODE_HEARTBEAT_GUARD << 7 | node_ID;
+	uint16_t COB_ID = FUNCTION_CODE_HEARTBEAT_GUARD << 7 | node_ID;
 	Hardware_CAN_Send_Message(COB_ID, data);
 }
 
@@ -26,6 +26,6 @@ void CANopen_HEARTBEAT_Protocol_Produce_Bootup_Mesage(CANopen *canopen, uint8_t 
 	uint8_t data[8] = {0};
 
 	/* Create the COB ID and send heartbeat to the node */
-	uint32_t COB_ID = FUNCTION_CODE_HEARTBEAT_GUARD << 7 | node_ID;
+	uint16_t COB_ID = FUNCTION_CODE_HEARTBEAT_GUARD << 7 | node_ID;
 	Hardware_CAN_Send_Message(COB_ID, data);
 }
