+++ src/Easy_CANopen/Easy_CANopen_Thread.c
@@ -13,7 +13,7 @@
 
 /* Call this function for every 1 millisecond */
 bool Easy_CANopen_Thread_Listen_Messages(CANopen *canopen) {
-	uint32_t COB_ID = 0;
+	uint16_t COB_ID = 0;
 	uint8_t data[8] = { 0 };
 	bool is_new_message = Hardware_CAN_Read_Message(&COB_ID, data);
 	if (is_new_message) {
